cleanup server_process to get rid of the most annoying warnings
The QProcess still gets destroyed too early at times. This lead to lots
of warnings, some of them duplicates and/or caused by other non-relevant
errors. Still haven't found a fix for the core issues but was able to
avoid them with a small standalone tool. TODO: find a real fix for this.