work on building protobuf as part of the project (WIP!)
The mess with protobuf, abseil, cmake, FindProtobuf and all that stuff is
just too much. It feels like there is no combination that works on most
distros, either cmake is too old or protobuf too new or something else
breaks. => Use cmakes FetchContent() to build protobuf as part of the
project.